--- .env
@@ -1,2 +1,0 @@
-KAKAO_NATIVE_APP_KEY=33ef3aa2164b50c8be59c1181ad95c41
-API_URL=https://junyeongan.store
