+++ .gitignore
@@ -45,3 +45,12 @@ app.*.map.json
 
 # FVM Version Cache
 .fvm/
+
+# Android Key
+/android/key.properties
+
+# Android Keystore
+/android/app/key.jks
+
+# Environment Variables
+.env
